feat: implemented redirect component
